Finished BRIQ-13 in 1.5 hours. Moved a part of BRIQ-11 to BRIQ-13, as it better suited BRIQ-13.
Instead of The player speed matching the speed cap in situations where the speed is calculated to be over it, it now decelerates to the speed cap. The side-ignite action has been changed to a straight-forward speed boost. Since the boost will normally put players over the speed cap, the speed gained will be lost quickly. The boost at the moment will go in the direction of the player's actual velocity. Will use play tests to determine if it should stay that way.